Service layer implementation

+++ src/main/java/com/bczovek/survey/api/model/SurveyStatistics.java
@@ -1,5 +1,5 @@
 package com.bczovek.survey.api.model;
 
-public record SurveyStatistics(Integer surveyId, String surveyName, Integer completedQuestionnaires,
-                               Integer filteredMembers, Integer rejectedMembers, Integer averageTime) {
+public record SurveyStatistics(Integer surveyId, String surveyName, Long completedQuestionnaires,
+                               Long filteredMembers, Long rejectedMembers, Double averageTime) {
 }
